Revert "EHN: Uvvis read transitions and MO contributions without pymatgen "
