+++ chemsmart/io/gaussian/output.py
@@ -1,18 +1,12 @@
-import re
-from itertools import islice
-from functools import cached_property
+from pymatgen.io.gaussian import GaussianOutput
 from chemsmart.utils.mixins import FileMixin
-
-
-# patterns for searching
-eV_pattern = r"([\d\.]+) eV"
-nm_pattern = r"([\d\.]+) nm"
-f_pattern = r"f=([\d\.]+)"
+from functools import cached_property
 
 
 class Gaussian16Output(FileMixin):
     def __init__(self, filename):
         self.filename = filename
+        self.gaussian_object = GaussianOutput(filename=filename)  # Gaussian output object from pymatgen
 
     @cached_property
     def tddft_transitions(self):
@@ -23,87 +17,14 @@ class Gaussian16Output(FileMixin):
             A list: A list of tuple for each transition such as
                     [(energie (eV), lambda (nm), oscillatory strength), ... ]
         """
-        tddft_transitions = []
-        for line in self.contents:
-            if line.startswith("Excited State"):
-                eV_match = re.search(eV_pattern, line)
-                nm_match = re.search(nm_pattern, line)
-                f_match = re.search(f_pattern, line)
-                if eV_match and nm_match and f_match:
-                    # Extract and convert the matched values to float
-                    excitation_energy_eV = float(eV_match.group(1))
-                    absorption_wavelength = float(nm_match.group(1))
-                    oscillatory_strength = float(f_match.group(1))
-
-                    tddft_transitions.append(
-                        (
-                            excitation_energy_eV,
-                            absorption_wavelength,
-                            oscillatory_strength,
-                        )
-                    )
-        return tddft_transitions
+        return self.gaussian_object.read_excitation_energies()
 
     @cached_property
     def excitation_energies_eV(self):
         """
-        Read TDDFT transitions and return the transition energies in eV as a list.
+        Read TDDFT transitions and return the transition energies in eV as a list
         """
         excitation_energies_eV = []
         for i in self.tddft_transitions:
             excitation_energies_eV.append(i[0])
         return excitation_energies_eV
-
-    @cached_property
-    def absorptions_in_nm(self):
-        """Read TDDFT transitions and return the absorbed wavelengths in nm as a list."""
-        absorptions_in_nm = []
-        for i in self.tddft_transitions:
-            absorptions_in_nm.append(i[1])
-        return absorptions_in_nm
-
-    @cached_property
-    def oscillatory_strengths(self):
-        """Read TDDFT transitions and return the oscillatory strengths as a list."""
-        oscillatory_strengths = []
-        for i in self.tddft_transitions:
-            oscillatory_strengths.append(i[2])
-        return oscillatory_strengths
-
-    @cached_property
-    def transitions(self):
-        """Read TDDFT transitions and return the MO transitions."""
-        transitions, _ = self._read_transitions_and_contribution_coefficients()
-        return transitions
-
-    @cached_property
-    def contribution_coefficients(self):
-        """Read MO contribution coefficients."""
-        _, cc = self._read_transitions_and_contribution_coefficients()
-        return cc
-
-    def _read_transitions_and_contribution_coefficients(self):
-        transitions = []
-        contribution_coefficients = []
-        for i, line in enumerate(self.contents):
-            if line.startswith("Excited State"):
-                each_state_transitions = []
-                each_state_contribution_coefficients = []
-                # parse the lines that follow until an empty line is encountered
-                j = 1
-                while len(self.contents[i + j]) != 0:
-                    line_element = self.contents[i + j].split()
-                    if len(line_element) <= 4:
-                        mo_transition = " ".join(
-                            list(islice(line_element, len(line_element) - 1))
-                        )
-                        contribution_coefficient = float(line_element[-1])
-                        each_state_transitions.append(mo_transition)
-                        each_state_contribution_coefficients.append(
-                            contribution_coefficient
-                        )
-                    j += 1
-                transitions.append(each_state_transitions)
-                contribution_coefficients.append(each_state_contribution_coefficients)
-
-        return transitions, contribution_coefficients
