+++ chemsmart/utils/periodictable.py
@@ -6,11 +6,7 @@ class PeriodicTable:
 
     def to_element(self, element_str):
         """Function to convert lower case element to element with proper caps in periodic table."""
-        return (
-            element_str.upper()
-            if len(element_str) == 1
-            else f"{element_str[0].upper()}{element_str[1:]}"
-        )
+        return element_str.upper() if len(element_str) == 1 else f"{element_str[0].upper()}{element_str[1:]}"
 
     def sorted_periodic_table_list(self, list_of_elements):
         return sorted(list_of_elements, key=lambda x: self.PERIODIC_TABLE.index(x))
